[DAY6-177] refactor : entity에서 DTO 변환을 service단으로 수정
