+++ src/main/java/day6/fullbang/controller/PlaceController.java
@@ -1,10 +1,7 @@
 package day6.fullbang.controller;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.List;
-import java.util.stream.Collectors;
 
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,10 +9,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import day6.fullbang.domain.Place;
 import day6.fullbang.domain.PlaceType;
-import day6.fullbang.domain.Product;
-import day6.fullbang.domain.Room;
 import day6.fullbang.dto.request.CoordinateRangeByDateDto;
 import day6.fullbang.dto.request.FilterOptionRequestDto;
 import day6.fullbang.dto.request.SearchRequestDto;
@@ -40,16 +34,9 @@ public class PlaceController {
             longitudeEnd,
             latitudeEnd, inputDate);
 
-        List<Place> placesByCoordinate = placeService.findPlacesByCoordinate(coordinateRangeByDateDto);
+        List<PlaceResponseDto> responseAllPlaces = placeService.findPlacesByCoordinate(coordinateRangeByDateDto);
 
-        List<PlaceResponseDto> responsePlaces = new ArrayList<>();
-        for (Place place : placesByCoordinate) {
-            Long lowestPrice = calculateLowestPriceInputDate(place.getRooms());
-            PlaceResponseDto item = new PlaceResponseDto(place, lowestPrice);
-            responsePlaces.add(item);
-        }
-
-        return responsePlaces;
+        return responseAllPlaces;
 
     }
 
@@ -70,12 +57,9 @@ public class PlaceController {
             new FilterOptionRequestDto(parkingAvailability, placeType, longitudeStart, latitudeStart, longitudeEnd,
                 latitudeEnd, maximumCapacity, inputDate);
 
-        List<Place> filteredPlaces = placeService.findPlacesByOption(filterOptionRequestDto);
-        List<PlaceResponseDto> filteredResponsePlaces = filteredPlaces.stream()
-            .map(p -> new PlaceResponseDto(p, calculateLowestPriceInputDate(p.getRooms())))
-            .collect(Collectors.toList());
+        List<PlaceResponseDto> responseFilteredPlaces = placeService.findPlacesByOption(filterOptionRequestDto);
 
-        return filteredResponsePlaces;
+        return responseFilteredPlaces;
     }
 
     @GetMapping("/search/{keyword}")
@@ -84,33 +68,10 @@ public class PlaceController {
         @RequestParam("longitude") Double longitude,
         @RequestParam("inputDate") @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate inputDate) {
         SearchRequestDto searchRequestDto = new SearchRequestDto(keyword, latitude, longitude, inputDate);
-        List<Place> placesByPlaceName = placeService.findPlacesByPlaceName(searchRequestDto);
 
-        List<PlaceResponseDto> responsePlaces = new ArrayList<>();
-        for (Place place : placesByPlaceName) {
-            Long lowestPrice = calculateLowestPriceInputDate(place.getRooms());
-            PlaceResponseDto item = new PlaceResponseDto(place, lowestPrice);
-            responsePlaces.add(item);
-        }
+        List<PlaceResponseDto> responseSearchPlaces = placeService.findPlacesByPlaceName(searchRequestDto);
 
-        return responsePlaces;
-    }
-
-    public Long calculateLowestPriceInputDate(List<Room> rooms) {
-        List<Long> priceList = new ArrayList<>();
-        // LocalDate nowDate = LocalDate.now(ZoneId.of("Asia/Seoul")); //현재 날짜
-
-        for (Room room : rooms) {
-            List<Product> products = room.getProducts();
-
-            for (Product product : products) {
-                if (product.getType().equals("숙박")) {
-                    priceList.add(product.getPrice());
-                }
-            }
-        }
-
-        return priceList.stream().min(Comparator.comparing(v -> v)).orElse(0L);
+        return responseSearchPlaces;
     }
 
 }
